// src/dynamic/directive/field-action-type.enum.ts
export enum FieldActionType {
    load = 'load',
    unload = 'unload',
    afterViewInit = 'afterViewInit',

    select = 'select',

    change = 'change',
    selectionChange = 'selectionChange',

    focus = 'focus',
    blur = 'blur',

    cut = 'cut',
    copy = 'copy',
    paste = 'paste',

    keydown = 'keydown',
    keypress = 'keypress',
    keyup = 'keyup',
    input = 'input',
    compositionend = 'compositionend',

    mouseenter = 'mouseenter',
    mouseleave = 'mouseleave',
    mousedown = 'mousedown',
    mouseup = 'mouseup',
    mousemove = 'mousemove',
    mouseover = 'mouseover',
    mouseout = 'mouseout',

    click = 'click',
    dblclick = 'dblclick',

    drag = 'drag',
    dragover = 'dragover',
    drop = 'drop',

    contextMenu = 'contextMenu',
    tabKeyDown = 'tabKeyDown'
}